% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{embed_medium_feed}
\alias{embed_medium_feed}
\title{Embed medium blog post as bootstrap carousel}
\usage{
embed_medium_feed(publication = "data-slice", tag = "")
}
\arguments{
\item{publication}{the name of the publication to embed}

\item{tag}{the tag to filter for}
}
\value{
an embedded medium feed
}
\description{
Embed medium publication blog post as a carousel of cards.
Carousel code inspired by bootstrap example with captions in
bootstrap documentation.
}
