% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{project_card}
\alias{project_card}
\title{Projects card}
\usage{
project_card(link, img, name, desc, img_alt = "")
}
\arguments{
\item{link}{the link to navigate to when clicking (e.g. project repo)}

\item{img}{the image to use}

\item{name}{name of the project}

\item{desc}{a short description of the project}
}
\value{
HTML code to provide a clickable image card for a project
}
\description{
Enter relevant info and links to create a card for a project with details etc.
}
