% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the app}
\usage{
run_app()
}
\value{
either run the app as a side effect or return a shiny.appobj object
}
\description{
Initiate an instance of the chrisbrownlie.com shiny app. This function will initiate the app stored in
inst/app with the runApp function
}
\details{
Note: an alias function 'l' is also provided to speed up workflow (after making changes, load them with 'Ctrl+Shift+L',
then launch the app with 'l()')
}
