% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{socials_bar}
\alias{socials_bar}
\title{Function to return a socials bar}
\usage{
socials_bar(links)
}
\arguments{
\item{links}{a list of lists with two named elements: link (the address
to link the icon to) and icon (the name of an icon)}
}
\value{
HTML for a socials bar on the main page of website
}
\description{
Enter links to your socials and return a bar
}
