% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{timeline}
\alias{timeline}
\title{Timeline}
\usage{
timeline(title, stuff)
}
\arguments{
\item{title}{the title of the timeline}

\item{stuff}{list, each element has date, title and info}
}
\value{
HTML code to provide a timeline of activities
}
\description{
Create a timeline for work experience info
}
